+++ quackosm/pbf_file_reader.py
@@ -32,6 +32,7 @@ import pyarrow.parquet as pq
 import shapely.wkt as wktlib
 from geoarrow.pyarrow import io
 from pandas.util._decorators import deprecate, deprecate_kwarg
+from pooch import get_logger as get_pooch_logger
 from pooch import retrieve
 from pooch.utils import parse_url
 from shapely.geometry import LinearRing, Polygon
@@ -698,6 +699,161 @@ class PbfFileReader:
 
         return gdf_parquet
 
+    def convert_pbf_to_duckdb(
+        self,
+        pbf_path: Union[str, Path, Iterable[Union[str, Path]]],
+        result_file_path: Optional[Union[str, Path]] = None,
+        keep_all_tags: bool = False,
+        explode_tags: Optional[bool] = None,
+        ignore_cache: bool = False,
+        filter_osm_ids: Optional[list[str]] = None,
+        duckdb_table_name: Optional[str] = "quackosm",
+    ) -> Path:
+        """
+        Convert PBF file to DuckDB Database.
+
+        Function parses multiple PBF files and returns a single GeoDataFrame with parsed
+        OSM objects.
+
+        Args:
+            pbf_path (Union[str, Path, Iterable[Union[str, Path]]]):
+                Path or list of paths of `*.osm.pbf` files to be parsed. Can be an URL.
+            result_file_path (Union[str, Path], optional): Where to save
+                the duckdb file. If not provided, will be generated based on hashes
+                from provided tags filter and geometry filter. Defaults to `None`.
+            keep_all_tags (bool, optional): Works only with the `tags_filter` parameter.
+                Whether to keep all tags related to the element, or return only those defined
+                in the `tags_filter`. When `True`, will override the optional grouping defined
+                in the `tags_filter`. Defaults to `False`.
+            explode_tags (bool, optional): Whether to split tags into columns based on OSM tag keys.
+                If `None`, will be set based on `tags_filter` and `keep_all_tags` parameters.
+                If there is tags filter defined and `keep_all_tags` is set to `False`, then it will
+                be set to `True`. Otherwise it will be set to `False`. Defaults to `None`.
+            ignore_cache: (bool, optional): Whether to ignore precalculated geoparquet files or not.
+                Defaults to False.
+            filter_osm_ids: (list[str], optional): List of OSM features ids to read from the file.
+                Have to be in the form of 'node/<id>', 'way/<id>' or 'relation/<id>'.
+                Defaults to an empty list.
+            duckdb_table_name (str): Table name in which data will be stored inside the DuckDB
+                database (default: "quackosm")
+
+        Returns:
+            gpd.GeoDataFrame: GeoDataFrame with OSM features.
+        """
+        if isinstance(pbf_path, (str, Path)):
+            pbf_path = [pbf_path]
+
+        parsed_geoparquet_file = self.convert_pbf_to_parquet(
+            pbf_path=pbf_path,
+            keep_all_tags=keep_all_tags,
+            explode_tags=explode_tags,
+            ignore_cache=ignore_cache,
+            filter_osm_ids=filter_osm_ids,
+        )
+
+        if filter_osm_ids is None:
+            filter_osm_ids = []
+
+        # generate result_file_path if missing
+        result_file_path = Path(
+            result_file_path
+            or self._generate_result_file_path(
+                pbf_path=pbf_path,
+                filter_osm_ids=filter_osm_ids,
+                keep_all_tags=keep_all_tags,
+                explode_tags=explode_tags or False,
+                save_as_wkt=False,
+            ).with_suffix(".duckdb")
+        )
+
+        result_file_path.parent.mkdir(exist_ok=True, parents=True)
+
+        duckdb_table_name = duckdb_table_name or "quackosm"
+
+        with duckdb.connect(str(result_file_path)) as con:
+            con.load_extension("spatial")
+            con.sql(f"""
+                CREATE OR REPLACE TABLE {duckdb_table_name} AS
+                SELECT * FROM read_parquet('{parsed_geoparquet_file}');
+            """)
+
+        # clean up intermediary parquet
+        parsed_geoparquet_file.unlink()
+
+        return result_file_path
+
+    def convert_geometry_to_duckdb(
+        self,
+        result_file_path: Optional[Union[str, Path]] = None,
+        keep_all_tags: bool = False,
+        explode_tags: Optional[bool] = None,
+        ignore_cache: bool = False,
+        filter_osm_ids: Optional[list[str]] = None,
+        duckdb_table_name: str = "quackosm",
+    ) -> Path:
+        """
+        Get features GeoDataFrame from a provided geometry filter.
+
+        Will automatically find and download OSM extracts covering a given geometry
+        and return a single GeoDataFrame with parsed OSM objects.
+
+        Args:
+            result_file_path (Union[str, Path], optional): Where to save
+                the duckdb file. If not provided, will be generated based on hashes
+                from provided tags filter and geometry filter. Defaults to `None`.
+            keep_all_tags (bool, optional): Works only with the `tags_filter` parameter.
+                Whether to keep all tags related to the element, or return only those defined
+                in the `tags_filter`. When `True`, will override the optional grouping defined
+                in the `tags_filter`. Defaults to `False`.
+            explode_tags (bool, optional): Whether to split tags into columns based on OSM tag keys.
+                If `None`, will be set based on `tags_filter` and `keep_all_tags` parameters.
+                If there is tags filter defined and `keep_all_tags` is set to `False`, then it will
+                be set to `True`. Otherwise it will be set to `False`. Defaults to `None`.
+            ignore_cache: (bool, optional): Whether to ignore precalculated geoparquet files or not.
+                Defaults to False.
+            filter_osm_ids: (list[str], optional): List of OSM features ids to read from the file.
+                Have to be in the form of 'node/<id>', 'way/<id>' or 'relation/<id>'.
+                Defaults to an empty list.
+            duckdb_table_name (str): Table name in which data will be stored inside the DuckDB
+                database (default: "quackosm")
+
+        Returns:
+            gpd.GeoDataFrame: GeoDataFrame with OSM features.
+        """
+        parsed_geoparquet_file = self.convert_geometry_to_parquet(
+            keep_all_tags=keep_all_tags,
+            explode_tags=explode_tags,
+            ignore_cache=ignore_cache,
+            filter_osm_ids=filter_osm_ids,
+        )
+
+        if filter_osm_ids is None:
+            filter_osm_ids = []
+
+        # generate result_file_path if missing
+        result_file_path = Path(
+            result_file_path
+            or self._generate_result_file_path_from_geometry(
+                filter_osm_ids=filter_osm_ids,
+                keep_all_tags=keep_all_tags,
+                explode_tags=explode_tags or False,
+                save_as_wkt=False,
+            ).with_suffix(".duckdb")
+        )
+
+        with duckdb.connect(str(result_file_path)) as con:
+            con.load_extension("spatial")
+
+            con.sql(f"""
+                CREATE OR REPLACE TABLE {duckdb_table_name} AS
+                SELECT * FROM read_parquet('{parsed_geoparquet_file}');
+            """)
+
+        # clean up intermediary parquet
+        parsed_geoparquet_file.unlink()
+
+        return result_file_path
+
     def _drop_duplicated_features_in_pyarrow_table(
         self, parsed_geoparquet_files: list[Path], tmp_dir_path: Path
     ) -> list[Path]:
@@ -792,6 +948,9 @@ class PbfFileReader:
         save_as_wkt: bool = False,
     ) -> Path:
         if _is_url_path(pbf_path):
+            logger = get_pooch_logger()
+            logger.setLevel("WARNING")
+
             pbf_path = retrieve(
                 pbf_path,
                 fname=Path(pbf_path).name,
@@ -911,7 +1070,7 @@ class PbfFileReader:
                 '{filtered_nodes_with_geometry_path}/**',
                 '{filtered_ways_with_proper_geometry_path}/**',
                 '{filtered_relations_with_geometry_path}/**'
-            ]);
+            ], union_by_name=True);
             """
         )
 
@@ -1597,6 +1756,11 @@ class PbfFileReader:
         self._run_query(query, run_in_separate_process)
         if self.debug_memory:
             log_message(f"Saved to directory: {file_path}")
+
+        is_empty = not any(file_path.iterdir())
+        if is_empty:
+            relation.to_parquet(str(file_path / "empty.parquet"))
+
         return self.connection.sql(f"SELECT * FROM read_parquet('{file_path}/**')")
 
     def _run_query(
@@ -1646,10 +1810,12 @@ class PbfFileReader:
         if result_path is None:
             result_path = file_path / "distinct"
 
+        relation = self.connection.sql(f"SELECT id FROM read_parquet('{file_path}/**') GROUP BY id")
+
         self.connection.sql(
             f"""
             COPY (
-                SELECT id FROM read_parquet('{file_path}/**') GROUP BY id
+                {relation.sql_query()}
             ) TO '{result_path}' (
                 FORMAT 'parquet',
                 PER_THREAD_OUTPUT true,
@@ -1660,6 +1826,10 @@ class PbfFileReader:
         )
         if self.debug_memory:
             log_message(f"Saved to directory: {result_path}")
+
+        is_empty = not any(result_path.iterdir())
+        if is_empty:
+            relation.to_parquet(str(result_path / "empty.parquet"))
 
         return self.connection.sql(f"SELECT * FROM read_parquet('{result_path}/**')")
 
@@ -2144,6 +2314,7 @@ class PbfFileReader:
             file_path=self.tmp_dir_path / "relation_inner_parts",
             step_name="Saving relations inner parts",
         )
+        any_relation_inner_parts = relation_inner_parts_parquet.count("id").fetchone()[0] != 0
         relation_outer_parts = self.connection.sql(
             f"""
             SELECT id, geometry_id, ST_MakePolygon(ST_RemoveRepeatedPoints(geometry)) geometry
@@ -2156,18 +2327,31 @@ class PbfFileReader:
             file_path=self.tmp_dir_path / "relation_outer_parts",
             step_name="Saving relations outer parts",
         )
-        relation_outer_parts_with_holes = self.connection.sql(
-            f"""
-            SELECT
-                og.id,
-                og.geometry_id,
-                ST_Difference(any_value(og.geometry), ST_Union_Agg(ig.geometry)) geometry
-            FROM ({relation_outer_parts_parquet.sql_query()}) og
-            JOIN ({relation_inner_parts_parquet.sql_query()}) ig
-            ON og.id = ig.id AND ST_WITHIN(ig.geometry, og.geometry)
-            GROUP BY og.id, og.geometry_id
-            """
-        )
+        if any_relation_inner_parts:
+            relation_outer_parts_with_holes = self.connection.sql(
+                f"""
+                SELECT
+                    og.id,
+                    og.geometry_id,
+                    ST_Difference(any_value(og.geometry), ST_Union_Agg(ig.geometry)) geometry
+                FROM ({relation_outer_parts_parquet.sql_query()}) og
+                JOIN ({relation_inner_parts_parquet.sql_query()}) ig
+                ON og.id = ig.id AND ST_WITHIN(ig.geometry, og.geometry)
+                GROUP BY og.id, og.geometry_id
+                """
+            )
+        else:
+            # Fake empty relation
+            relation_outer_parts_with_holes = self.connection.sql(
+                f"""
+                SELECT
+                    og.id,
+                    og.geometry_id,
+                    og.geometry
+                FROM ({relation_outer_parts_parquet.sql_query()}) og
+                WHERE og.id IS NULL
+                """
+            )
         relation_outer_parts_with_holes_parquet = self._save_parquet_file_with_geometry(
             relation=relation_outer_parts_with_holes,
             file_path=self.tmp_dir_path / "relation_outer_parts_with_holes",
@@ -2230,7 +2414,7 @@ class PbfFileReader:
                 f"""
                 COPY (
                     SELECT
-                        * EXCLUDE (geometry), ST_AsWKB(ST_MakeValid(geometry)) geometry_wkb
+                        * EXCLUDE (geometry), ST_MakeValid(geometry::GEOMETRY) geometry
                     FROM ({relation.sql_query()})
                 ) TO '{file_path}' (
                     FORMAT 'parquet',
@@ -2243,9 +2427,17 @@ class PbfFileReader:
             if self.debug_memory:
                 log_message(f"Saved to directory: {file_path}")
 
+        is_empty = not any(file_path.iterdir())
+        if is_empty:
+            relation.to_parquet(str(file_path / "empty.parquet"))
+
         return self.connection.sql(
             f"""
-            SELECT * EXCLUDE (geometry_wkb), ST_GeomFromWKB(geometry_wkb) geometry
+            SELECT * EXCLUDE(geometry),
+            CASE WHEN typeof(geometry) = 'GEOMETRY'
+            THEN geometry::GEOMETRY
+            ELSE ST_GeomFromWKB(geometry::BLOB)
+            END AS geometry
             FROM read_parquet('{file_path}/**')
             """
         )
@@ -2263,7 +2455,7 @@ class PbfFileReader:
             *self._generate_osm_tags_sql_select(
                 parsed_geometries, keep_all_tags=keep_all_tags, explode_tags=explode_tags
             ),
-            "ST_GeomFromWKB(geometry_wkb) AS geometry",
+            GEOMETRY_COLUMN,
         ]
 
         unioned_features = self.connection.sql(
@@ -2277,18 +2469,9 @@ class PbfFileReader:
             unioned_features, keep_all_tags=keep_all_tags, explode_tags=explode_tags
         )
 
-        features_full_relation = self.connection.sql(
-            f"""
-            SELECT
-                * EXCLUDE (geometry),
-                ST_MakeValid(geometry) geometry
-            FROM ({grouped_features.sql_query()})
-            """
-        )
-
         features_parquet_path = self.tmp_dir_path / "osm_valid_elements"
         self._save_parquet_file_with_geometry(
-            features_full_relation,
+            grouped_features,
             features_parquet_path,
             step_name="Saving all features",
         )
@@ -2563,9 +2746,11 @@ class PbfFileReader:
                 else:
                     geometry_column = ga.as_wkb(gpd.GeoSeries([], crs=WGS84_CRS))
 
-                features_parquet_table = features_parquet_table.append_column(
-                    GEOMETRY_COLUMN, geometry_column
-                ).drop("geometry_wkb")
+                features_parquet_table = features_parquet_table.set_column(
+                    features_parquet_table.schema.get_field_index(GEOMETRY_COLUMN),
+                    GEOMETRY_COLUMN,
+                    geometry_column,
+                )
 
                 features_parquet_table = features_parquet_table.select(
                     [FEATURES_INDEX, GEOMETRY_COLUMN]
@@ -2583,7 +2768,7 @@ class PbfFileReader:
             columns_to_test = [
                 f'COUNT_IF("{col}" IS NOT NULL) == 0 as "{col}"'
                 for col in features_table.columns
-                if col not in (FEATURES_INDEX, "geometry_wkb")
+                if col not in (FEATURES_INDEX, GEOMETRY_COLUMN)
             ]
             columns_to_test_result = self.connection.sql(
                 f"SELECT {', '.join(columns_to_test)} FROM '{input_file}/*.parquet'"
@@ -2609,17 +2794,21 @@ class PbfFileReader:
 
                         if save_as_wkt:
                             geometry_column = ga.as_wkt(
-                                ga.with_crs(batch.column("geometry_wkb"), WGS84_CRS)
+                                ga.with_crs(batch.column(GEOMETRY_COLUMN), WGS84_CRS)
                             )
-                            batch = batch.drop_columns("geometry_wkb").append_column(
-                                GEOMETRY_COLUMN, geometry_column
+                            batch = batch.set_column(
+                                batch.schema.get_field_index(GEOMETRY_COLUMN),
+                                GEOMETRY_COLUMN,
+                                geometry_column,
                             )
                         else:
                             geometry_column = ga.as_wkb(
-                                ga.with_crs(batch.column("geometry_wkb"), WGS84_CRS)
+                                ga.with_crs(batch.column(GEOMETRY_COLUMN), WGS84_CRS)
                             )
-                            batch = batch.drop_columns("geometry_wkb").append_column(
-                                GEOMETRY_COLUMN, geometry_column
+                            batch = batch.set_column(
+                                batch.schema.get_field_index(GEOMETRY_COLUMN),
+                                GEOMETRY_COLUMN,
+                                geometry_column,
                             )
                             batch = _replace_geo_metadata_in_batch(batch)
 
@@ -2664,15 +2853,19 @@ class PbfFileReader:
                         geometry_column = ga.as_wkt(
                             ga.with_crs(batch.column(GEOMETRY_COLUMN), WGS84_CRS)
                         )
-                        batch = batch.drop_columns(GEOMETRY_COLUMN).append_column(
-                            GEOMETRY_COLUMN, geometry_column
+                        batch = batch.set_column(
+                            batch.schema.get_field_index(GEOMETRY_COLUMN),
+                            GEOMETRY_COLUMN,
+                            geometry_column,
                         )
                     else:
                         geometry_column = ga.as_wkb(
                             ga.with_crs(batch.column(GEOMETRY_COLUMN), WGS84_CRS)
                         )
-                        batch = batch.drop_columns(GEOMETRY_COLUMN).append_column(
-                            GEOMETRY_COLUMN, geometry_column
+                        batch = batch.set_column(
+                            batch.schema.get_field_index(GEOMETRY_COLUMN),
+                            GEOMETRY_COLUMN,
+                            geometry_column,
                         )
                         batch = _replace_geo_metadata_in_batch(batch)
 
@@ -2710,8 +2903,7 @@ def _set_up_duckdb_connection(
     connection.sql("SET enable_progress_bar_print = false;")
 
     connection.install_extension("spatial")
-    for extension_name in ("parquet", "spatial"):
-        connection.load_extension(extension_name)
+    connection.load_extension("spatial")
 
     connection.sql(
         """