+++ tests/base/test_cli.py
@@ -104,6 +104,12 @@ def test_transient_mode(monaco_pbf_file_path_fixture: str) -> None:
 @P.case("Ignore cache short", ["--no-cache"], "files/monaco_nofilter_noclip_compact.parquet")  # type: ignore
 @P.case("Output", ["--output", "files/monaco_output.parquet"], "files/monaco_output.parquet")  # type: ignore
 @P.case("Output short", ["-o", "files/monaco_output.parquet"], "files/monaco_output.parquet")  # type: ignore
+@P.case("DuckDB explicit export", ["--duckdb"], "files/monaco_nofilter_noclip_compact.duckdb")  # type: ignore
+@P.case(
+    "DuckDB explicit export with table name",
+    ["--duckdb", "--duckdb-table-name", "test"],
+    "files/monaco_nofilter_noclip_compact.duckdb",
+)  # type: ignore
 @P.case("Silent", ["--silent"], "files/monaco_nofilter_noclip_compact.parquet")  # type: ignore
 @P.case("Transient", ["--transient"], "files/monaco_nofilter_noclip_compact.parquet")  # type: ignore
 @P.case(
@@ -275,6 +281,11 @@ def test_proper_args_with_pbf(
 
 @P.parameters("args", "expected_result")  # type: ignore
 @P.case(
+    "Geometry BBOX filter",
+    ["--geom-filter-bbox", "7.416486,43.731086,7.421931,43.733707"],
+    "files/b9115f99abe0ba4933e368fdf5e7d4ffbcc6fa48ba5ce03c4c50f95165fd3291_nofilter_compact.parquet",
+)  # type: ignore
+@P.case(
     "Geometry WKT filter",
     ["--geom-filter-wkt", geometry_wkt()],
     "files/09c3fc0471538594b784be7c52782837c7a26753c2b26097b780581fa0a6bfc6_nofilter_compact.parquet",
@@ -335,14 +346,20 @@ def test_proper_args_with_pbf(
     "files/6a869bcfa1a49ade8b76569e48e4142bce29098815bf37e57155a18204f2bbbc_nofilter_compact.parquet",
 )  # type: ignore
 @P.case(
-    "Working directory",
-    ["--geom-filter-file", geometry_boundary_file_path(), "--working-directory", "files/workdir"],
-    "files/workdir/6a869bcfa1a49ade8b76569e48e4142bce29098815bf37e57155a18204f2bbbc_nofilter_compact.parquet",
-)  # type: ignore
-@P.case(
     "Ignore cache",
     ["--geom-filter-file", geometry_boundary_file_path(), "--ignore-cache"],
     "files/6a869bcfa1a49ade8b76569e48e4142bce29098815bf37e57155a18204f2bbbc_nofilter_compact.parquet",
+)  # type: ignore
+@P.case(
+    "Working directory",
+    [
+        "--geom-filter-file",
+        geometry_boundary_file_path(),
+        "--working-directory",
+        "files/workdir",
+        "--ignore-cache",
+    ],
+    "files/workdir/6a869bcfa1a49ade8b76569e48e4142bce29098815bf37e57155a18204f2bbbc_nofilter_compact.parquet",
 )  # type: ignore
 @P.case(
     "Output",
@@ -633,6 +650,8 @@ def test_proper_args_with_pbf_url() -> None:
 @P.case("Geometry H3 filter with S2", ["--geom-filter-index-h3", "12cdc28bc"])  # type: ignore
 @P.case("Geometry H3 filter with random string", ["--geom-filter-index-h3", random_str()])  # type: ignore
 @P.case("Geometry S2 filter with random string", ["--geom-filter-index-s2", random_str()])  # type: ignore
+@P.case("Geometry BBOX filter with wrong values", ["--geom-filter-bbox", random_str()])  # type: ignore
+@P.case("Geometry BBOX filter with spaces", ["--geom-filter-bbox", "10,", "-5,", "6,", "17"])  # type: ignore
 @P.case(
     "Geometry two filters",
     ["--geom-filter-geojson", geometry_geojson(), "--geom-filter-wkt", geometry_wkt()],
@@ -676,6 +695,14 @@ def test_proper_args_with_pbf_url() -> None:
     [
         "--osm-extract-query",
         "nonexistent_extract",
+    ],
+)  # type: ignore
+@P.case(
+    "OSM extracts with zero matches and duckdb export",
+    [
+        "--duckdb",
+        "--osm-extract-query",
+        "quack_extract",
     ],
 )  # type: ignore
 @P.case(