+++ tests/benchmark/test_big_file.py
@@ -4,6 +4,7 @@ import shutil
 from pathlib import Path
 
 from parametrization import Parametrization as P
+from pooch import get_logger as get_pooch_logger
 from pooch import retrieve
 
 from quackosm import PbfFileReader, geocode_to_geometry
@@ -22,6 +23,8 @@ def test_big_file(extract_name: str, geocode_filter: list[str], tags_filter: Osm
     files_dir = Path("files")
     shutil.rmtree(files_dir)
     file_name = files_dir / f"{extract_name}.osm.pbf"
+    logger = get_pooch_logger()
+    logger.setLevel("WARNING")
     retrieve(
         f"https://download.geofabrik.de/europe/{extract_name}-latest.osm.pbf",
         fname=f"{extract_name}.osm.pbf",